Remove API Keys page and update settings UI
- Deleted the APIKeysPage component to streamline API key management.
- Updated the router to remove the API Keys route and adjusted the SettingsLayout to reflect this change.
- Modified the description in the SettingsLayout for the API section to clarify its purpose.
- Enhanced the APISettingsPage to incorporate API key management functionalities directly, improving user experience and accessibility.
